Task view revamp; added processing status on main view

+++ src/main/java/kw/tools/gallery/views/MainView.java
@@ -1,5 +1,6 @@
 package kw.tools.gallery.views;
 
+import com.vaadin.flow.component.Component;
 import com.vaadin.flow.component.Key;
 import com.vaadin.flow.component.UI;
 import com.vaadin.flow.component.Unit;
@@ -7,17 +8,21 @@ import com.vaadin.flow.component.button.Button;
 import com.vaadin.flow.component.grid.Grid;
 import com.vaadin.flow.component.grid.GridVariant;
 import com.vaadin.flow.component.html.Anchor;
+import com.vaadin.flow.component.html.Span;
 import com.vaadin.flow.component.icon.VaadinIcon;
 import com.vaadin.flow.component.orderedlayout.HorizontalLayout;
 import com.vaadin.flow.component.orderedlayout.VerticalLayout;
 import com.vaadin.flow.component.page.Page;
+import com.vaadin.flow.component.progressbar.ProgressBar;
 import com.vaadin.flow.component.textfield.TextField;
 import com.vaadin.flow.data.renderer.ComponentRenderer;
 import com.vaadin.flow.data.renderer.TextRenderer;
 import com.vaadin.flow.router.Route;
 import kw.tools.gallery.models.Repository;
 import kw.tools.gallery.services.GalleryService;
+import kw.tools.gallery.services.ProcessingService;
 import kw.tools.gallery.services.RepositoryService;
+import kw.tools.gallery.services.TaskService;
 import org.springframework.beans.factory.annotation.Autowired;
 
 @Route("")
@@ -27,15 +32,23 @@ public class MainView extends VerticalLayout
 
     private final GalleryService galleryService;
 
+    private final ProcessingService processingService;
+
+    private final TaskService taskService;
+
     private final Grid<Repository> repositoryGrid = new Grid<>();
     private final TextField repositoryInput = new TextField();
 
 
     public MainView(@Autowired RepositoryService repositoryService,
-                    @Autowired GalleryService galleryService)
+                    @Autowired GalleryService galleryService,
+                    @Autowired ProcessingService processingService,
+                    @Autowired TaskService taskService)
     {
         this.repositoryService = repositoryService;
         this.galleryService = galleryService;
+        this.processingService = processingService;
+        this.taskService = taskService;
 
         setHeight(100, Unit.PERCENTAGE);
 
@@ -73,6 +86,21 @@ public class MainView extends VerticalLayout
         repositoryGrid.addColumn(new TextRenderer<>(repository -> Integer.toString(galleryService.getGalleryCountForRepo(repository.getId()))))
                 .setKey("galCount").setHeader("Galleries").setSortable(true).setAutoWidth(true)
                 .setFlexGrow(0).setResizable(true);
+
+        repositoryGrid.addComponentColumn(repository -> {
+            VerticalLayout wrapper = new VerticalLayout();
+            wrapper.add(new Anchor("tasks/" + repository.getId(), interpretProcessingStatus(processingService.getProcessingStatus(repository.getId()))));
+            if (!interpretIsProcessingFinished(processingService.getProcessingStatus(repository.getId())))
+            {
+                ProgressBar pg = new ProgressBar();
+                pg.setMin(0);
+                pg.setMax(galleryService.getGalleryCountForRepo(repository.getId()));
+                pg.setValue(processingService.getFinishedCount(repository.getId()));
+                wrapper.add(pg);
+            }
+            wrapper.setAlignItems(Alignment.CENTER);
+            return wrapper;
+        }).setHeader("Status").setFlexGrow(0).setResizable(false);
 
         repositoryGrid.addColumn(new ComponentRenderer<>(repo -> {
             Button refresh = new Button();
@@ -114,13 +142,40 @@ public class MainView extends VerticalLayout
     private void deleteRepo(String id)
     {
         Page page = UI.getCurrent().getPage();
-        page.executeJs(String.format("return confirm(\"Are you sure you want to delete '%s'?\")", id)).then(result -> {
+        page.executeJs(String.format("return confirm(\"Are you sure you want to delete '%s' with all its galleries and data?\")", id)).then(result -> {
             if (result.asBoolean())
             {
                 repositoryService.delete(id);
                 galleryService.deleteAll(id);
+                taskService.clearByCategory(id);
                 refreshGrid();
             }
         });
+    }
+
+    private Component interpretProcessingStatus(ProcessingService.ProcessingStatus status)
+    {
+        switch (status)
+        {
+            case SUCCESSFUL:
+                return VaadinIcon.CHECK.create();
+            case ERRORS:
+                return VaadinIcon.WARNING.create();
+            case IDLE:
+                return VaadinIcon.QUESTION.create();
+            case WORKING:
+                return VaadinIcon.HOURGLASS.create();
+            case UNEXPECTED:
+                return new Span("Unexpected status");
+            default:
+                return new Span("Status not implemented!");
+        }
+    }
+
+    private boolean interpretIsProcessingFinished(ProcessingService.ProcessingStatus status)
+    {
+        return status == ProcessingService.ProcessingStatus.SUCCESSFUL
+                || status == ProcessingService.ProcessingStatus.ERRORS
+                || status == ProcessingService.ProcessingStatus.UNEXPECTED;
     }
 }
